doit: Fix issues identified by -Wextra
At least make everything compile. Future work is to clean up all the
legitimate unused parameters that we seem to have (?!).

+++ ilpc.h
@@ -6,6 +6,7 @@
 
 #include <stddef.h>
 #include <stdint.h>
+#include <sys/types.h>
 
 #include "sio.h"
 
@@ -19,8 +20,8 @@ int ilpcb_destroy(struct ilpcb *ctx);
 int ilpcb_probe(struct ilpcb *ctx);
 
 /* These are going to be *amazingly* slow. Use the l2ab */
-int ilpcb_read(struct ilpcb *ctx, size_t addr, void *buf, size_t len);
-int ilpcb_write(struct ilpcb *ctx, size_t addr, const void *buf, size_t len);
+ssize_t ilpcb_read(struct ilpcb *ctx, size_t addr, void *buf, size_t len);
+ssize_t ilpcb_write(struct ilpcb *ctx, size_t addr, const void *buf, size_t len);
 
 int ilpcb_readl(struct ilpcb *ctx, size_t addr, uint32_t *val);
 int ilpcb_writel(struct ilpcb *ctx, size_t addr, uint32_t val);
